rag-backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.59.3
supabase==2.11.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
tiktoken==0.8.0
langchain-text-splitters==0.3.4
python-multipart==0.0.9
